requirements.txt
torch
diffusers==0.16.1
transformers
scipy
ftfy
accelerate
tqdm
requests
fastapi
Pillow==9.5.0
aiogram==2.23.1
pyngrok
uvicorn
python-multipart
